Add stock controller

+++ src/main/java/com/dbms/backend/repo/employee/EmployeeRepo.java
@@ -22,31 +22,31 @@ public class EmployeeRepo {
     public EmployeeRepo(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
-    public List<EmployeeDetails> getEmployeeDetails() {
+    public List<EmployeeDetails> getEmployeeDetails(int user_id) {
         try{
-            String sql = "SELECT * FROM employee";
-            return jdbcTemplate.query(sql, new EmployeeDetailsRowMapper());
+            String sql = "SELECT * FROM employee WHERE user_id = ?";
+            return jdbcTemplate.query(sql, new EmployeeDetailsRowMapper(), user_id);
         }catch (Exception e) {
             throw new RuntimeException(e);
         }
     }
 
-    public void addEmployee(EmployeeDetails employeeDetails) {
+    public void addEmployee(EmployeeDetails employeeDetails, int user_id) {
         try{
-            String sql = "INSERT INTO employee (name, role, address, join_date, salary) VALUES (?, ?, ?, ?, ?)";
-            jdbcTemplate.update(sql, employeeDetails.name(), employeeDetails.role(), employeeDetails.address(), employeeDetails.join_date(), employeeDetails.salary());
+            String sql = "INSERT INTO employee (name, role, address, join_date, salary, user_id) VALUES (?, ?, ?, ?, ?, ?)";
+            jdbcTemplate.update(sql, employeeDetails.name(), employeeDetails.role(), employeeDetails.address(), employeeDetails.join_date(), employeeDetails.salary(), user_id);
         }catch (Exception e) {
             throw new RuntimeException(e);
         }
     }
 
-    public Optional<EmployeeCompleteDetails> getEmployeeDetailsById(int id){
+    public Optional<EmployeeCompleteDetails> getEmployeeDetailsById(int id, int user_id){
         try{
-            String sql = "SELECT * FROM employee WHERE id = ?";
+            String sql = "SELECT * FROM employee WHERE id = ? AND user_id = ?";
             String sqlPhones = "SELECT * FROM employee_phone WHERE employee_id = ?";
             String sqlEmails = "SELECT * FROM employee_email WHERE employee_id = ?";
             String sqlAttendance = "SELECT * FROM employee_leave WHERE employee_id = ?";
-            EmployeeDetails employeeDetails = jdbcTemplate.query(sql, new EmployeeDetailsRowMapper(), id).stream().findFirst().orElse(null);
+            EmployeeDetails employeeDetails = jdbcTemplate.query(sql, new EmployeeDetailsRowMapper(), id, user_id).stream().findFirst().orElse(null);
             if(employeeDetails == null){
                 return Optional.empty();
             }
@@ -62,26 +62,27 @@ public class EmployeeRepo {
         }
     }
     @Transactional
-    public void deleteEmployeeById(int id) {
+    public void deleteEmployeeById(int id, int user_id) {
         try{
-            String sql = "DELETE FROM employee WHERE id = ?";
+            String sql = "DELETE FROM employee WHERE id = ? AND user_id = ?";
             String sqlPhones = "DELETE FROM employee_phone WHERE employee_id = ?";
             String sqlEmails = "DELETE FROM employee_email WHERE employee_id = ?";
             String sqlDates = "DELETE FROM employee_leave WHERE employee_id = ?";
             // Deleting Phone numbers, emails and leave dates of the employee when he gets deleted
+            // TODO: Need to check if the employee exist or not --> For validation
             jdbcTemplate.update(sqlPhones, id);
             jdbcTemplate.update(sqlEmails, id);
             jdbcTemplate.update(sqlDates, id);
-            jdbcTemplate.update(sql, id);
+            jdbcTemplate.update(sql, id, user_id);
         }catch (Exception e) {
             throw new RuntimeException(e);
         }
     }
 
-    public void updateEmployeeDetailsById(int id, EmployeeDetails employeeDetails) {
+    public void updateEmployeeDetailsById(int id, EmployeeDetails employeeDetails, int user_id) {
         try{
-            String sql = "UPDATE employee SET name = ?, role = ?, address = ?, join_date = ?, salary = ? WHERE id = ?";
-            jdbcTemplate.update(sql, employeeDetails.name(), employeeDetails.role(), employeeDetails.address(), employeeDetails.join_date(), employeeDetails.salary(), id);
+            String sql = "UPDATE employee SET name = ?, role = ?, address = ?, join_date = ?, salary = ? WHERE id = ? AND user_id = ?";
+            jdbcTemplate.update(sql, employeeDetails.name(), employeeDetails.role(), employeeDetails.address(), employeeDetails.join_date(), employeeDetails.salary(), id, user_id);
         }catch (Exception e) {
             throw new RuntimeException(e);
         }
